requirements.txt
python-dotenv
requests
beautifulsoup4
supabase
langgraph
langchain-openai
langchain-core
pydantic
typing-extensions
bert-score
rouge
